requirements.txt
ezdxf==1.4.3
fonttools==4.61.1
lxml==6.0.2
numpy==2.4.1
pillow==12.1.0
pyparsing==3.3.2
typing_extensions==4.15.0
pydantic-settings==2.6.1
loguru==0.7.2
cos-python-sdk-v5==1.9.36
python-dotenv==1.0.1
pika==1.3.2
requests==2.32.3
